+++ ChangeLog.md
@@ -1,3 +1,8 @@
-# Changelog for haskeleton
+# Changelog for water-cooler
+
+## 0.1.1.0
+ * The water-cooler `next` command now prints time until the next drink in "human-friendly" time.
+   I.e. instead of "120s" it will print "2m".
+   Currently it uses the floor calculation; therefore "119s" would show as "1m"
 
 ## Unreleased changes